refactor: input 컴포넌트 파일 경로 변경
