+++ src/stories/Input/ExpirationDateInput.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { ExpirationDateInput } from '../../components/ExpirationDateInput';
+import { ExpirationDateInput } from '../../components/input/ExpirationDateInput';
 
 const meta = {
   title: 'Example/Input',
